still working on mealDB favorites, the image isn't loading when selected

+++ Meal Ideas/Screens/MealDB/MealDB Detail View/MealDBDetailVM.swift
@@ -5,34 +5,50 @@
 //  Created by Steve Plavetzky on 11/30/21.
 //
 
-import Foundation
+import UIKit
 import CoreData
 
 @MainActor final class MealDBDetailVM: ObservableObject{
-    @Published var meal: MealDBResults.Meal
+    @Published var meal: MealDBResults.Meal?
     @Published var isLoading = false
     @Published var favorited : Bool
+    @Published var mealID: String
+    @Published var mealPhoto = UIImage()
+
     
-    init(meal : MealDBResults.Meal, favorited: Bool){
+    init(meal : MealDBResults.Meal?, favorited: Bool, mealID: String){
         self.meal = meal
         self.favorited = favorited
+        self.mealID = mealID
         fetchMeal()
     }
     // MARK: - Fetch Meal
     func fetchMeal(){
-        let mealID = meal.id
-        print("Fetching MealDB Single Named mealID: \(mealID ?? "")")
-        isLoading = false
-        Task {
-            do {
-                let results = try await NetworkManager.shared.mealDBQuery(query: mealID ?? "", queryType: .none)
-                if let safeResults = results.first{
-                    meal = safeResults
+        if meal == nil || meal?.ingredientsArray == []{
+            print("No Meal Provided, Fetching MealDB Single Named mealID: \(mealID)")
+            isLoading = false
+            Task {
+                do {
+                    let results = try await NetworkManager.shared.mealDBQuery(query: mealID, queryType: .none)
+                    if let safeResults = results.first{
+                        meal = safeResults
+                    }
                 }
             }
         }
     }
     
+    // MARK: - Get Meal Photo
+    func getMealPhoto(){
+        if meal != nil{
+            NetworkManager.shared.downloadImage(fromURLString: meal?.strMealThumb ?? "") { uiImage in
+                guard let uiImage = uiImage else { return }
+                DispatchQueue.main.async {
+                    self.mealPhoto = uiImage
+                }
+            }
+        }
+    }
     /*
     // MARK: - Get Favorites
     func getFavorites(){
@@ -66,22 +82,25 @@ import CoreData
 */
     // MARK: - Favorite Toggled
     func favoriteToggled(){
-        if favorited == true {
-            //add to favorites
-            print("add to favorite")
-            PersistenceController.shared.saveFavorites(mealName: meal.strMeal ?? "",
-                                mealDBID: meal.id,
-                                spoonID: nil)
-            
-        } else {
-            //remove from favorite
-            print("remove from favorites")
-            
-            PersistenceController.shared.deleteFavorite(source: .mealDB,
-                                                        mealName: meal.strMeal ?? "",
-                                                        mealDBID: meal.id,
-                                                        spoonID: nil)
+        if let safeMeal = meal{
+            if favorited == true {
+                //add to favorites
+                print("add to favorite")
+                PersistenceController.shared.saveFavorites(mealName: safeMeal.strMeal ?? "",
+                                    mealDBID: safeMeal.id,
+                                    spoonID: nil)
+                
+            } else {
+                //remove from favorite
+                print("remove from favorites")
+                
+                PersistenceController.shared.deleteFavorite(source: .mealDB,
+                                                            mealName: safeMeal.strMeal ?? "",
+                                                            mealDBID: safeMeal.id,
+                                                            spoonID: nil)
+            }
         }
+
     }
 }
